WIP: more tests

+++ fonts/fonts.go
@@ -115,10 +115,10 @@ type FontMetrics interface {
 	// should return a default value (the upem number for example).
 	// If the glyph is invalid it should return 0.
 	// `coords` is used by variable fonts, and is specified in normalized coordinates.
-	GetHorizontalAdvance(gid GlyphIndex, coords []float32) int16
+	GetHorizontalAdvance(gid GlyphIndex, coords []float32) float32
 
 	// Same as `GetHorizontalAdvance`, but for vertical advance.
-	GetVerticalAdvance(gid GlyphIndex, coords []float32) int16
+	GetVerticalAdvance(gid GlyphIndex, coords []float32) float32
 
 	// Fetches the (X,Y) coordinates of the origin (in font units) for a glyph ID,
 	// for horizontal text segments.
